Run Sphinx doctests with Nose
Added requirements for testing. Also fixed a test code block ambiguity
that causes doctests fail in Nose due to different test ordering.

Sphinx's doctest module implement its own test runner; the runner is
quite simple, and runs doctests sequentially, as they appear in docs.
Nose, on the other hand, is based on the built-in test suite runner,
which run tests *in arbitrary order*. This is problematic if you have
doctest blocks like this:

    >>> a = 'foo'    # Block 1

    >>> a = 'bar'    # Block 2

    >>> print(a)     # Block 3

In Sphinx, tests run sequentially, so block 3 is guarenteed to output
"bar". In Nose, however, they can run in any order, so block 3 might
output "foo" (2-1-3), "bar" (1-2-3). It can even raise a NameError
(if run before both blocks 1 and 2).

This is solved by declaring those lines as an explicit doctest block.

    .. doctest::

        >>> a = 'foo'    # Block 1

        >>> a = 'bar'    # Block 2

        >>> print(a)     # Block 3

This way, Sa'im al-Dahr will collect them as one single test, thus
the order of statements can be guarenteed.

Tests using dict are moved to unittests and converted to use
OrderedDict so that sanity of tests is promised. Print statements
are converted to print functions a la Python 3 (via __future__).