Minor optimisations in util and autoparam rewrite
Perform string coersion only on Python 2, and use version-dependent definitions on
Python 2 and 3 for qualifier and _is_iterable_not_str.

autoparam is now a seperate constant. Passing the param class directly will not
work now; use the mosql.util.autoparam token instead. ___ is now a shortcut to
autoparam.

Implementation of __repr__ fixed for param and raw classes.

Docs and tests modified to reflect the change.